consumer_api/requirements.txt
Flask>=1.1.1
Flask-Cors>=3.0.8
python-dotenv>=0.10.3
confluent-kafka>=1.3.0
requests>=2.22.0
pytest>=5.1.2
flask-swagger-ui>=3.20.9
gunicorn==20.0.0
gevent==1.4.0